#15[Feat] 카테고리별 영상 칩과 연결완료
