+++ app/src/main/java/com/brandon/playvideo_app/data/model/CategoryVideoModel.kt
@@ -1,4 +1,4 @@
-package com.brandon.playvideo_app.data.category
+package com.brandon.playvideo_app.data.model
 
 import com.google.gson.annotations.SerializedName
