--- app/src/main/java/com/brandon/playvideo_app/data/trendvideo/Default.kt
@@ -1,12 +1,0 @@
-package com.brandon.playvideo_app.data.trendvideo
-
-import com.google.gson.annotations.SerializedName
-
-data class Default(
-    @SerializedName("height")
-    val height: Int,
-    @SerializedName("url")
-    val url: String,
-    @SerializedName("width")
-    val width: Int
-)
