--- app/src/main/java/com/brandon/playvideo_app/data/trendvideo/Item.kt
@@ -1,14 +1,0 @@
-package com.brandon.playvideo_app.data.trendvideo
-
-import com.google.gson.annotations.SerializedName
-
-data class Item(
-    @SerializedName("etag")
-    val etag: String,
-    @SerializedName("id")
-    val id: String,
-    @SerializedName("kind")
-    val kind: String,
-    @SerializedName("snippet")
-    val snippet: Snippet
-)
