--- app/src/main/java/com/brandon/playvideo_app/data/trendvideo/PageInfo.kt
@@ -1,10 +1,0 @@
-package com.brandon.playvideo_app.data.trendvideo
-
-import com.google.gson.annotations.SerializedName
-
-data class PageInfo(
-    @SerializedName("resultsPerPage")
-    val resultsPerPage: Int,
-    @SerializedName("totalResults")
-    val totalResults: Int
-)
