--- app/src/main/java/com/brandon/playvideo_app/data/trendvideo/Thumbnails.kt
@@ -1,16 +1,0 @@
-package com.brandon.playvideo_app.data.trendvideo
-
-import com.google.gson.annotations.SerializedName
-
-data class Thumbnails(
-    @SerializedName("default")
-    val default: Default,
-    @SerializedName("high")
-    val high: High,
-    @SerializedName("maxres")
-    val maxres: Maxres,
-    @SerializedName("medium")
-    val medium: Medium,
-    @SerializedName("standard")
-    val standard: Standard
-)
