--- app/src/main/java/com/brandon/playvideo_app/data/trendvideo/TrendVideoModel.kt
@@ -1,16 +1,0 @@
-package com.brandon.playvideo_app.data.trendvideo
-
-import com.google.gson.annotations.SerializedName
-
-data class TrendVideoModel(
-    @SerializedName("etag")
-    val etag: String,
-    @SerializedName("items")
-    val items: List<Item>,
-    @SerializedName("kind")
-    val kind: String,
-    @SerializedName("nextPageToken")
-    val nextPageToken: String,
-    @SerializedName("pageInfo")
-    val pageInfo: PageInfo
-)
