+++ app/src/main/res/layout/category_fragment.xml
@@ -10,6 +10,7 @@
         android:id="@+id/horizontal_scrollview"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_marginHorizontal="10dp"
         android:scrollbars="none"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
