+++ app/src/main/res/layout/item_category.xml
@@ -1,17 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content">
 
     <ImageView
         android:id="@+id/iv_category_thumbnail"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginHorizontal="10dp"
         android:adjustViewBounds="true"
-        android:minWidth="180dp"
-        android:minHeight="220dp"
+        android:maxHeight="240dp"
+        android:maxWidth="220dp"
         android:scaleType="fitXY"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -20,12 +21,15 @@
     <TextView
         android:id="@+id/tv_category_title"
         style="@style/video_title_style"
-        android:layout_width="match_parent"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="12dp"
-        tools:text="Love wins all (Love wins all) "
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/iv_category_thumbnail" />
+        android:ellipsize="end"
+        android:maxLines="2"
+        app:layout_constraintEnd_toEndOf="@id/iv_category_thumbnail"
+        app:layout_constraintStart_toStartOf="@id/iv_category_thumbnail"
+        app:layout_constraintTop_toBottomOf="@id/iv_category_thumbnail"
+        tools:text="Love wins all (Love wins all) " />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
